+++ demo/example.js
@@ -64,20 +64,3 @@ export const home = `
   </Actions>
 </Home>
 `.trim()
-
-export const dialog = `
-<!-- ⚠️ Please notice that dialog components were deprecated. -->
-<Dialog callbackId="createUser" title="Create user">
-  <Input name="name" label="Name" required />
-  <Textarea name="desc" label="Description" maxLength="300" />
-
-  <Select name="role" label="Role" value="regular" required>
-    <Option value="regular">Regular</Option>
-    <Option value="leader">Leader</Option>
-    <Option value="admin">Admin</Option>
-  </Select>
-
-  <Input type="hidden" name="userId" value="xxxxxxxx" />
-  <Input type="submit" value="Create" />
-</Dialog>
-`.trim()